+++ lib/ruby_smart/support/gem_version.rb
@@ -9,7 +9,7 @@ module RubySmart
 
     module VERSION
       MAJOR = 1
-      MINOR = 2
+      MINOR = 3
       TINY  = 0
       PRE   = nil